+++ react-components/src/api.js
@@ -87,16 +87,16 @@ export default {
     return get(config.apiSuggestedCountriesUrl, { hs_code }).then((response) => responseHandler(response).json())
   },
 
-  getCountryData: (country) => {
-    return get(config.countryDataUrl, { country: country }).then((response) => responseHandler(response).json())
-  },
-
   getPopulationByCountryData: (countries) => {
     return get(config.populationByCountryUrl, { countries: countries }).then((response) => responseHandler(response).json())
   },
 
   getComTradeData: (countries, commodity_code) => {
-  return get(config.apiComTradeDataUrl, { countries: countries, commodity_code: commodity_code }).then((response) => responseHandler(response).json())
+    return get(config.apiComTradeDataUrl, { countries: countries, commodity_code: commodity_code }).then((response) => responseHandler(response).json())
+  },
+
+  getCountryAgeGroupData: (data) => {
+    return get(config.countryAgeGroupDataUrl, data).then((response) => responseHandler(response).json())
   },
 
   getMarketingCountryData: (data) => {
@@ -137,6 +137,18 @@ export default {
 
   updateRouteToMarket: (data) => {
     return post(config.apiRouteToMarketUpdateUrl, data).then((response) => responseHandler(response).json())
+  },
+
+  createAdaptTarketMarketDocumentList: (data) => {
+    return post(config.apiTargetMarketDocumentsCreateUrl, data).then((response) => responseHandler(response).json())
+  },
+
+  deleteAdaptTarketMarketDocumentList: (pk) => {
+    return httpDelete(config.apiTargetMarketDocumentsDeleteUrl, { pk: pk }).then(responseHandler)
+  },
+
+  updateAdaptTarketMarketDocumentList: (data) => {
+    return post(config.apiTargetMarketDocumentsUpdateUrl, data).then((response) => responseHandler(response).json())
   },
 
   createObjective: (data) => {