+++ react-components/src/components/ProductFinder/ProductFinder.test.jsx
@@ -189,7 +189,7 @@ describe('Product finder tests', () => {
       let results = finder.querySelector('.scroll-area div');
       expect(results).toBeTruthy()
     })
-    expect(finder.querySelector('.box').textContent).toMatch(/^The item you are classifying is considered a complex item/)
+    expect(finder.querySelector('.form-group-error').textContent).toMatch(/We couldn't find a match because your search had too many product names/)
   })
 
   it('Opens product view and renames product', async () => {
@@ -224,7 +224,7 @@ describe('Product finder tests', () => {
       expect(results).toBeTruthy()
     })
     const box = finder.querySelector('.box');
-    expect(box.querySelector('h3').textContent).toMatch('HS Code: 123456')
+    expect(box.querySelector('h3').textContent).toMatch('HS6 Code: 123456')
     expect(box.querySelector('input').getAttribute('value')).toMatch('my product')
   })