+++ react-components/src/components/ProductFinder/StartEndPage.test.jsx
@@ -3,7 +3,8 @@ import React, { useState } from 'react'
 import ReactDOM from 'react-dom'
 import fetchMock from 'fetch-mock'
 import Services from '@src/Services'
-import { waitFor } from '@testing-library/react'
+import { waitFor, fireEvent } from '@testing-library/react'
+
 import { act, Simulate } from 'react-dom/test-utils'
 import StartEndPage from './StartEndPage'
 
@@ -29,7 +30,7 @@ const mockResponse = {
 it('Renders a product selection final page', async () => {
   const hsCode = '123456'
   const commodityName = "default name"
-  const newName = "updated name"
+  const newName = " updated name "
   const saveProduct = jest.fn()
 
   Services.setConfig({ apiLookupProductScheduleUrl: '/api/lookup-product-schedule/' })
@@ -39,17 +40,26 @@ it('Renders a product selection final page', async () => {
       ReactDOM.render(
       <StartEndPage 
         commodityCode={hsCode} 
-        defaultCommodityName="default name" 
+        defaultCommodityName={commodityName} 
         saveProduct={saveProduct} 
       />, 
       container)
   })
   await waitFor(() => {
-    let results = container.querySelector('input');
+    let results = container.querySelector('input')
     expect(results).toBeTruthy()
   })
   let nameInput = container.querySelector('input')
   expect(nameInput.getAttribute('value')).toEqual(commodityName)
+  // to test the clear button - we need to focus the input.
+  // We have to change the content and fire a change event to make that happen as
+  // focus event doesn't fire on the window in test
+  nameInput.focus()
+  act(() => {
+    nameInput.value=commodityName + '-'
+    Simulate.change(nameInput)
+  })
+  
   const clearButton = container.querySelector('button.clear')
   act(() => {
     Simulate.click(clearButton)
@@ -68,6 +78,6 @@ it('Renders a product selection final page', async () => {
   })
   expect(saveProduct.mock.calls.length).toEqual(1)
   expect(saveProduct.mock.calls[0][0]).toEqual(hsCode)
-  expect(saveProduct.mock.calls[0][1]).toEqual(newName)
-
+  expect(saveProduct.mock.calls[0][1]).not.toEqual(newName)
+  expect(saveProduct.mock.calls[0][1]).toEqual(newName.trim())
 })